+++ src/pages/technologies/list.tsx
@@ -1,10 +1,26 @@
-import { keepPreviousData, useQuery } from "@tanstack/react-query";
+import { IconCircleCheckFilled, IconDotsVertical, IconLoader } from "@tabler/icons-react";
+import { keepPreviousData, useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
+import { ColumnDef } from "@tanstack/react-table";
 import { useCallback, useState } from "react";
 import { useAuth } from "react-oidc-context";
+import { Link } from "react-router";
 import { toast } from "sonner";
+import { z } from "zod";
+
+import { Badge } from "@/ui/badge";
+import { Button } from "@/ui/button";
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuSeparator,
+  DropdownMenuTrigger,
+} from "@/ui/dropdown-menu";
 
 import { API_URL } from "@/constants/application";
-import { GET_TECHNOLOGIES } from "@/constants/query-keys";
+import { DELETE_TECHNOLOGY, GET_TECHNOLOGIES } from "@/constants/query-keys";
+
+import { technologySchema } from "@/schemas/technology";
 
 import { createQueryParams } from "@/lib/params";
 
@@ -12,12 +28,83 @@ import { TechnologyTable } from "@/components/technologies/table";
 
 export const TechnologiesPage = () => {
   const auth = useAuth();
+  const queryClient = useQueryClient();
   const [queryParams, setQueryParams] = useState({
     page: 1,
     size: 10,
     sort: ["title", "asc"],
   });
 
+  const columns: ColumnDef<z.infer<typeof technologySchema>>[] = [
+    {
+      accessorKey: "title",
+      header: "Title",
+      cell: ({ row }) => <div className="w-32">{row.original.title}</div>,
+      enableHiding: false,
+    },
+    {
+      accessorKey: "website",
+      header: "Website",
+      cell: ({ row }) => <div className="w-32">{row.original.website}</div>,
+    },
+    {
+      accessorKey: "moved",
+      header: "Moved",
+      cell: ({ row }) => (
+        <div className="w-32">
+          <Badge variant="outline" className="text-muted-foreground px-1.5">
+            {row.original.moved}
+          </Badge>
+        </div>
+      ),
+      enableColumnFilter: false,
+    },
+    {
+      accessorKey: "active",
+      header: "Active",
+      cell: ({ row }) => (
+        <Badge variant="outline" className="text-muted-foreground px-1.5">
+          {row.original.active ? (
+            <IconCircleCheckFilled className="fill-green-500 dark:fill-green-400" />
+          ) : (
+            <IconLoader />
+          )}
+          {row.original.active}
+        </Badge>
+      ),
+      enableColumnFilter: false,
+    },
+  ];
+
+  const columnsTech: ColumnDef<z.infer<typeof technologySchema>>[] = [
+    ...columns,
+    {
+      id: "actions",
+      cell: ({ row }) => (
+        <DropdownMenu>
+          <DropdownMenuTrigger asChild>
+            <Button
+              variant="ghost"
+              className="data-[state=open]:bg-muted text-muted-foreground flex size-8"
+              size="icon"
+            >
+              <IconDotsVertical />
+              <span className="sr-only">Open menu</span>
+            </Button>
+          </DropdownMenuTrigger>
+          <DropdownMenuContent align="end" className="w-32">
+            <Link to={`/technologies/edit/${row.id}`}>
+              <DropdownMenuItem className="cursor-pointer">Edit</DropdownMenuItem>
+            </Link>
+            <DropdownMenuSeparator />
+            <DropdownMenuItem variant="destructive" onClick={() => deleteTechnology(row.id)}>
+              Delete
+            </DropdownMenuItem>
+          </DropdownMenuContent>
+        </DropdownMenu>
+      ),
+    },
+  ];
   const {
     data: technologiesData = { content: [], pageable: { pageNumber: 0, pageSize: 10 }, totalElements: 0 },
     isLoading: isFetchingTechnologiesData,
@@ -36,6 +123,28 @@ export const TechnologiesPage = () => {
       return await response.json();
     },
     placeholderData: keepPreviousData,
+  });
+
+  const { mutate: deleteTechnology } = useMutation({
+    mutationFn: async (rowId: string) => {
+      await fetch(`${API_URL}/technologies/${rowId}`, {
+        method: "DELETE",
+        headers: {
+          "Content-type": "application/json",
+          Authorization: `Bearer ${auth.user?.access_token}`,
+        },
+      });
+    },
+    mutationKey: [DELETE_TECHNOLOGY],
+    onSuccess() {
+      queryClient.invalidateQueries({ queryKey: [GET_TECHNOLOGIES] });
+      toast.success("Technology has been deleted successfully");
+    },
+    onError(error) {
+      toast.error("Error deleting technology", {
+        description: error.message,
+      });
+    },
   });
 
   if (isErrorDataList) {
@@ -64,11 +173,13 @@ export const TechnologiesPage = () => {
     <>
       <TechnologyTable
         data={technologiesData.content}
+        columns={columnsTech}
         handlePagination={handlePaginationParams}
-        rowCount={technologiesData.totalElements}
-        isLoading={isFetchingTechnologiesData}
         handleSorting={handleSortingParams}
         handleFilter={handleFilterParams}
+        handleDelete={deleteTechnology}
+        rowCount={technologiesData.totalElements}
+        isLoading={isFetchingTechnologiesData}
         pageSize={technologiesData.pageable.pageSize}
         pageIndex={technologiesData.pageable.pageNumber}
       />
